Update for userCosin function
Finally got the data how I wanted to read it. Now figuring out the best way to store this data whether in a 2d array or just traverse through a slice with all the data in it. Need to figure out the best approach so can best execute cosine similarity.